More splitting of roundtrip tests into partial files
