+++ src/Kingsland.MofParser.UnitTests/CodeGen/RoundtripTests_ClassDeclaration.cs
@@ -1,9 +1,4 @@
-﻿using Kingsland.MofParser.CodeGen;
-using Kingsland.MofParser.Lexing;
-using Kingsland.MofParser.Parsing;
-using Kingsland.ParseFx.Parsing;
-using Kingsland.ParseFx.Text;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
 namespace Kingsland.MofParser.UnitTests.CodeGen
 {
@@ -19,125 +14,64 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
             [Test]
             public static void EmptyClassDeclarationShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Base\r\n" +
                     "{\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void ClassDeclarationWithSuperclassShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Base : GOLF_Superclass\r\n" +
                     "{\r\n" +
                     "\tstring InstanceID;\r\n" +
                     "\tstring Caption = Null;\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void ClassDeclarationWithClassFeaturesShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Base\r\n" +
                     "{\r\n" +
                     "\tstring InstanceID;\r\n" +
                     "\tstring Caption = Null;\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void ClassDeclarationsWithQualifierListShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "[Abstract, OCL{\"-- the key property cannot be NULL\", \"inv: InstanceId.size() = 10\"}]\r\n" +
                     "class GOLF_Base\r\n" +
                     "{\r\n" +
                     "\t[Description(\"an instance of a class that derives from the GOLF_Base class. \"), Key] string InstanceID;\r\n" +
                     "\t[Description(\"A short textual description (one- line string) of the\"), MaxLen(64)] string Caption = Null;\r\n" +
-                    "};"
-                );
-            }
-
-        }
-
-        public static class ClassFeatureTests
-        {
-
-            [Test]
-            public static void ClassFeatureWithQualifiersShouldRoundtrip()
-            {
-                RoundtripTests.AssertRoundtrip(
-                    "class Sponsor\r\n" +
-                    "{\r\n" +
-                    "\t[Description(\"Monthly salary in $US\")] string Name;\r\n" +
-                    "};"
-                );
-            }
-
-            [Test]
-            public static void InvalidClassFeatureShouldThrow()
-            {
-                var sourceMof =
-                    "class Sponsor\r\n" +
-                    "{\r\n" +
-                    "\t100\r\n" +
                     "};";
-                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
-                var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
-                var ex = Assert.Throws<UnexpectedTokenException>(
-                    () => {
-                        var astNodes = Parser.Parse(tokens);
-                    }
-                );
-                Assert.AreEqual(
-                    "Unexpected token found at Position 19, Line Number 3, Column Number 2.\r\n" +
-                    "Token Type: 'IntegerLiteralToken'\r\n" +
-                    "Token Text: '100'",
-                    ex.Message
-                );
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
-            [Test]
-            public static void ClassFeatureWithStructureDeclarationShouldRoundtrip()
-            {
-                RoundtripTests.AssertRoundtrip(
-                    "class Sponsor\r\n" +
-                    "{\r\n" +
-                    "\tstructure Nested\r\n" +
-                    "\t{\r\n" +
-                    "\t};\r\n" +
-                    "};"
-                );
-            }
-
-            [Test]
-            public static void ClassFeatureWithEnumerationDeclarationShouldRoundtrip()
-            {
-                RoundtripTests.AssertRoundtrip(
-                    "class Sponsor\r\n" +
-                    "{\r\n" +
-                    "\tenumeration MonthsEnum : Integer\r\n" +
-                    "\t{\r\n" +
-                    "\t};\r\n" +
-                    "};"
-                );
-            }
-
-            [Test]
-            public static void ClassFeatureWithPropertyDeclarationShouldRoundtrip()
-            {
-                RoundtripTests.AssertRoundtrip(
-                    "class Sponsor\r\n" +
-                    "{\r\n" +
-                    "\tstring Name;\r\n" +
-                    "};"
-                );
-            }
+            //[Test]
+            //public static void ClassDeclarationsAstWithNumericPropertiesShouldRoundtrip()
+            //{
+            //    RoundtripTests.AssertRoundtrip(
+            //        "instance of myType as $Alias00000070\r\n" +
+            //        "{\r\n" +
+            //        "\tMyBinaryValue = 0101010b;\r\n" +
+            //        "\tMyOctalValue = 0444444;\r\n" +
+            //        "\tMyHexValue = 0xABC123;\r\n" +
+            //        "\tMyDecimalValue = 12345;\r\n" +
+            //        "\tMyRealValue = 123.45;\r\n" +
+            //        "};"
+            //    );
+            //}
 
         }
