+++ src/Kingsland.MofParser.UnitTests/CodeGen/RoundtripTests_MethodDeclaration.cs
@@ -14,84 +14,84 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
             [Test]
             public static void MethodDeclarationWithNoParametersShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Club\r\n" +
                     "{\r\n" +
                     "\tInteger GetMembersWithOutstandingFees();\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void MethodDeclarationWithParameterShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Club\r\n" +
                     "{\r\n" +
                     "\tInteger GetMembersWithOutstandingFees(GOLF_ClubMember lateMembers);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void MethodDeclarationWithArrayParameterShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Club\r\n" +
                     "{\r\n" +
                     "\tInteger GetMembersWithOutstandingFees(GOLF_ClubMember lateMembers[]);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test]
             public static void MethodDeclarationsWithRefParameterShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Club\r\n" +
                     "{\r\n" +
                     "\tInteger GetMembersWithOutstandingFees(GOLF_ClubMember REF lateMembers);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test(Description = "https://github.com/mikeclayton/MofParser/issues/27")]
             public static void ClassDeclarationsWithMethodDeclarationWithEnumParameterShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Professional : GOLF_ClubMember\r\n" +
                     "{\r\n" +
                     "\tGOLF_ResultCodeEnum GetNumberOfProfessionals(Integer NoOfPros, GOLF_Club Club, ProfessionalStatusEnum Status = Professional);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test(Description = "https://github.com/mikeclayton/MofParser/issues/37")]
             public static void MethodDeclarationsWithArrayReturnTypeShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Club\r\n" +
                     "{\r\n" +
                     "\tInteger[] GetMembersWithOutstandingFees(GOLF_ClubMember lateMembers);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test(Description = "https://github.com/mikeclayton/MofParser/issues/38")]
             public static void MethodDeclarationWithMultipleParametersShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class GOLF_Professional : GOLF_ClubMember\r\n" +
                     "{\r\n" +
                     "\tGOLF_ResultCodeEnum GetNumberOfProfessionals(Integer NoOfPros, GOLF_Club Club, ProfessionalStatusEnum Status = Professional);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test(Description = "https://github.com/mikeclayton/MofParser/issues/28")]
             public static void MethodDeclarationWithDeprecatedMof300IntegerReturnTypesAndQuirksDisabledShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class Win32_SoftwareFeature : CIM_SoftwareFeature\r\n" +
                     "{\r\n" +
                     "\tuint8 ReinstallUint8(integer ReinstallMode = 1);\r\n" +
@@ -102,14 +102,14 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "\tsint16 ReinstallUint16(integer ReinstallMode = 1);\r\n" +
                     "\tsint32 ReinstallUint32(integer ReinstallMode = 1);\r\n" +
                     "\tsint64 ReinstallUint64(integer ReinstallMode = 1);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
             [Test(Description = "https://github.com/mikeclayton/MofParser/issues/28")]
             public static void MethodDeclarationWithDeprecatedMof300IntegerParameterTypesShouldRoundtrip()
             {
-                RoundtripTests.AssertRoundtrip(
+                var sourceText =
                     "class Win32_SoftwareFeature : CIM_SoftwareFeature\r\n" +
                     "{\r\n" +
                     "\tinteger ReinstallUint8(uint8 ReinstallMode = 1);\r\n" +
@@ -120,8 +120,8 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "\tinteger ReinstallUint16(sint16 ReinstallMode = 1);\r\n" +
                     "\tinteger ReinstallUint32(sint32 ReinstallMode = 1);\r\n" +
                     "\tinteger ReinstallUint64(sint64 ReinstallMode = 1);\r\n" +
-                    "};"
-                );
+                    "};";
+                RoundtripTests.AssertRoundtrip(sourceText);
             }
 
         }
